pyproject.toml
[tool.poetry]
name = "airflow_dev"
version = "0.1.0"
description = "Env dev"
authors = ["lcorrea"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
requests = "^2.31.0"
pymongo = "^4.6.0"
apache-airflow = "^2.7.3"
apache-airflow-providers-mongo = "^3.4.0"
pandas = "^2.1.3"
[tool.poetry.group.test.dependencies]  # https://python-poetry.org/docs/master/managing-dependencies/
black = { extras = ["jupyter"], version = ">=23.3.0" }
commitizen = ">=3.2.1"
coverage = { extras = ["toml"], version = ">=7.2.5" }
mypy = ">=1.2.0"
poethepoet = ">=0.20.0"
pre-commit = ">=3.3.1"
pytest = ">=7.3.1"
pytest-clarity = ">=1.0.1"
pytest-mock = ">=3.10.0"
pytest-xdist = ">=3.2.1"
ruff = ">=0.0.270"
safety = ">=2.3.4,!=2.3.5"
shellcheck-py = ">=0.9.0"
typeguard = ">=3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
